Stock module finished
